feat: extract mirrors data function
