+++ src/app/(search)/mirrors/page.tsx
@@ -16,9 +16,7 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
-import {fetchMirrorlist} from '@/lib/github';
-import {fetchRepoTimestamp} from '@/lib/mirrors';
-import {Mirror, RepoCheck, RepoStatus} from '@/lib/types';
+import {getMirrorsData} from '@/lib/mirrors';
 
 export const metadata: Metadata = {
   title: 'Mirrors List',
@@ -63,134 +61,7 @@ export default function Mirrors() {
   );
 }
 
-const PRIMARY_MIRROR_URL = 'https://build.cachyos.org/repo';
-const FETCH_TIMEOUT_MS = 6000;
-const SYNC_TOLERANCE_SECONDS = 3600;
-
-const REPO_PATHS = [
-  'x86_64/cachyos',
-  'x86_64_v3/cachyos-v3',
-  'x86_64_v3/cachyos-core-v3',
-  'x86_64_v3/cachyos-extra-v3',
-  'x86_64_v4/cachyos-v4',
-  'x86_64_v4/cachyos-core-v4',
-  'x86_64_v4/cachyos-extra-v4',
-  'x86_64_v4/cachyos-znver4',
-  'x86_64_v4/cachyos-core-znver4',
-  'x86_64_v4/cachyos-extra-znver4',
-] as const;
-
 async function MirrorslistPage() {
-  const mirrorsList = await fetchMirrorlist();
-
-  const baselineController = new AbortController();
-  const baselineTimeout = setTimeout(
-    () => baselineController.abort(),
-    FETCH_TIMEOUT_MS
-  );
-
-  const baselinePromises = REPO_PATHS.map(async path => ({
-    path,
-    timestamp: await fetchRepoTimestamp(
-      PRIMARY_MIRROR_URL,
-      path,
-      baselineController.signal
-    ),
-  }));
-
-  const baselines = await Promise.all(baselinePromises);
-  clearTimeout(baselineTimeout);
-
-  const baselineMap = new Map(baselines.map(b => [b.path, b.timestamp]));
-
-  const mirrorChecks = mirrorsList.map(async mirrorUrl => {
-    const controller = new AbortController();
-    const timeout = setTimeout(() => controller.abort(), FETCH_TIMEOUT_MS);
-
-    const checks = await Promise.all(
-      REPO_PATHS.map(async path => {
-        const timestamp = await fetchRepoTimestamp(
-          mirrorUrl,
-          path,
-          controller.signal
-        );
-        const baseline = baselineMap.get(path);
-
-        let status: RepoStatus = 'error';
-        let lag: null | number = null;
-
-        if (timestamp !== null) {
-          if (baseline) {
-            lag = baseline - timestamp;
-            status = lag <= SYNC_TOLERANCE_SECONDS ? 'synced' : 'out-of-sync';
-          } else {
-            status = 'synced';
-          }
-        }
-
-        return {
-          lastUpdated: timestamp,
-          path,
-          status,
-          syncLagSeconds: lag,
-        } satisfies RepoCheck;
-      })
-    );
-
-    clearTimeout(timeout);
-
-    const validChecks = checks.filter(c => c.status !== 'error');
-    const totalChecks = checks.length;
-    const errorChecks = checks.length - validChecks.length;
-    const syncedChecks = checks.filter(c => c.status === 'synced').length;
-
-    let overallStatus: Mirror['overallStatus'] = 'error';
-
-    if (validChecks.length === 0) {
-      overallStatus = 'error';
-    } else if (syncedChecks === totalChecks) {
-      overallStatus = 'healthy';
-    } else if (errorChecks > 0 || syncedChecks < totalChecks) {
-      overallStatus = 'partial';
-      if (syncedChecks === 0) overallStatus = 'out-of-sync';
-    }
-
-    const lags = validChecks
-      .map(c => c.syncLagSeconds)
-      .filter(l => l !== null)
-      .filter(l => l > 0);
-
-    const averageLag =
-      lags.length > 0 ? lags.reduce((a, b) => a + b, 0) / lags.length : null;
-
-    const url = new URL(mirrorUrl);
-
-    return {
-      averageLagSeconds: averageLag,
-      checks,
-      name: url.hostname,
-      overallStatus,
-      url: mirrorUrl,
-    } satisfies Mirror;
-  });
-
-  const mirrors = await Promise.all(mirrorChecks);
-
-  mirrors.sort((a, b) => {
-    const score = (s: Mirror['overallStatus']) => {
-      switch (s) {
-        case 'error':
-          return 3;
-        case 'healthy':
-          return 0;
-        case 'out-of-sync':
-          return 2;
-        case 'partial':
-          return 1;
-      }
-    };
-    return score(a.overallStatus) - score(b.overallStatus);
-  });
-
+  const {baselines, mirrors} = await getMirrorsData();
   return <MirrorslistTable baselines={baselines} mirrors={mirrors} />;
 }
